+++ src/pylmrob/__init__.py
@@ -13,13 +13,16 @@ from pylmrob.anova import anova
 from pylmrob.bootstrap import bootstrap
 from pylmrob.control import Control
 from pylmrob.lmrob import LmRob, lmrob
+from pylmrob.rng import RState, r_set_seed
 
 __all__ = [
     "Control",
     "LmRob",
+    "RState",
     "__version__",
     "anova",
     "bootstrap",
     "lmrob",
     "make_generator",
+    "r_set_seed",
 ]